moving dependecies to its own package
