+++ dependencies/gitreader.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package utils
+package dependencies
 
 import (
 	"archive/zip"
